requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
torchaudio 
transformers 
segment-anything-py
rembg
numpy<2
opencv-python
Pillow 
onnxruntime
requests
matplotlib